Add conformation when deleting or demoting a user

+++ src/pages/UserList/UserCard.jsx
@@ -3,14 +3,22 @@ import { deleteUser, demoteTrainer } from "../../services/auth";
 /* eslint-disable react/prop-types */
 export default function UserCard({ user, refreshUsers, setActiveUser }) {
     const onDemote = async () => {
-        await demoteTrainer(user._id);
-        await refreshUsers();
+        const result = confirm(`Are you sure you want to demote user ${user.username}?`);
+
+        if (result) {
+            await demoteTrainer(user._id);
+            await refreshUsers();
+        }
     }
 
     const onDelete = async () => {
-        await deleteUser(user._id);
-        await refreshUsers();
-        setActiveUser(() => { });
+        const result = confirm(`Are you sure you want to delete user ${user.username}?`);
+
+        if (result) {
+            await deleteUser(user._id);
+            await refreshUsers();
+            setActiveUser(() => { });
+        }
     }
